Fix Android compilation errors preventing APK build
Two compilation errors were preventing the build:

1. WeatherRepository.kt: WbgtCalculator is a Kotlin `object` (singleton)
   but was being instantiated with WbgtCalculator(). Removed the illegal
   constructor call and use the object directly.

2. ForecastScreen.kt: humidity (Double 65.0) was passed to
   HourlyForecast.humidity which expects Int. Added .toInt() conversion.

Also:
- Added ProGuard keep rule for API response classes (data.api package)
- Added --stacktrace to Gradle CI commands for better error diagnosis
- Used glob pattern for APK artifact upload paths

+++ Android/app/src/main/java/com/heatshield/agri/ui/screens/ForecastScreen.kt
@@ -51,7 +51,7 @@ fun ForecastScreen() {
                 hour = hour,
                 wbgt = result.wbgt,
                 temperature = temp,
-                humidity = humidity,
+                humidity = humidity.toInt(),
                 windSpeed = wind,
                 solarRadiation = solar
             )
